feat: Shared validation schema using pick()
