+++ apps/editor.planx.uk/src/pages/FlowEditor/components/Settings/Team/Design/ButtonColour/index.tsx
@@ -12,9 +12,8 @@ import SettingsFormContainer from "../../../shared/SettingsForm";
 import { GET_TEAM_THEME, UPDATE_TEAM_THEME } from "../shared/queries";
 import { DesignPreview } from "../shared/styles";
 import type { GetTeamTheme } from "../shared/types";
-import { defaultValues } from "../shared/validationSchema";
+import { defaultValues, validationSchema } from "../shared/validationSchema";
 import type { FormValues, MutationVars } from "./types";
-import { validationSchema } from "./validationSchema";
 
 const ButtonColour: React.FC = () => {
   const teamId = useStore((state) => state.teamId);
@@ -33,7 +32,7 @@ const ButtonColour: React.FC = () => {
           action_colour: theme.actionColour,
         },
       })}
-      validationSchema={validationSchema}
+      validationSchema={validationSchema.pick(["actionColour"])}
       legend="Button colour"
       preview={(formik) => [
         <DesignPreview bgcolor="white">
