--- apps/editor.planx.uk/src/pages/FlowEditor/components/Settings/Team/Design/ButtonColour/validationSchema.ts
@@ -1,5 +1,0 @@
-import { object, string } from "yup";
-
-export const validationSchema = object({
-  actionColour: string().required(),
-});
