--- apps/editor.planx.uk/src/pages/FlowEditor/components/Settings/Team/Design/Favicon/validationSchema.ts
@@ -1,5 +1,0 @@
-import { object, string } from "yup";
-
-export const validationSchema = object({
-  favicon: string().defined().nullable(),
-});
